requirements.txt
groq>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
markdown>=3.5.0

